// src/api-gen/todo/models/todo-create-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TodoCreateDto {

  /**
   * the description of the todo
   */
  description: string;

  /**
   * the unique key as GUID of the todo
   */
  guid: string;

  /**
   * the name of the todo
   */
  name: string;
}