// src/api-gen/todo/models/todo-return-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TodoReturnDto {

  /**
   * the field describes if the todo item is an active one.
   */
  active: boolean;

  /**
   * the description of the todo
   */
  description: string;

  /**
   * the unique key as GUID of the todo
   */
  guid: string;

  /**
   * the name of the todo
   */
  name: string;

  /**
   * is one of the states from the enum below
   */
  state: 'open' | 'closed';
}